Refactor: move impls from `protocol` to `protocol::types`
